Migrate memmem function to util.h/c
It is unclear to me why the module cannot call anything in libkern.h
despite including it. The module does compile but the symbol cannot be
found at load time. This is out of my depth and care so I have simply
copied the function to util.c for use.